refactor: use ducks pattern
