+++ front/src/modules/toastNotification.ts
@@ -13,8 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { Dispatch } from "redux";
-import actionCreatorFactory from "typescript-fsa";
+import { Action, Dispatch } from "redux";
+import actionCreatorFactory, { isType } from "typescript-fsa";
+
+export interface IState {
+  readonly notifications: Array<{
+    readonly isShown: boolean;
+    readonly message: string;
+    readonly kind: "success" | "error";
+    readonly id: number;
+  }>;
+}
 
 const actionCreator = actionCreatorFactory();
 
@@ -42,3 +51,39 @@ export const toastr = (
     }, 200);
   }, duration);
 };
+
+const INIT: IState = {
+  notifications: []
+};
+
+export function reducer(state: IState = INIT, action: Action): IState {
+  if (isType(action, showToastr)) {
+    return {
+      ...state,
+      notifications: [
+        {
+          isShown: true,
+          message: action.payload.message,
+          kind: action.payload.kind,
+          id: action.payload.id
+        },
+        ...state.notifications
+      ]
+    };
+  }
+  if (isType(action, hideToastr)) {
+    return {
+      ...state,
+      notifications: state.notifications.map(x =>
+        x.id === action.payload.id ? { ...x, isShown: false } : x
+      )
+    };
+  }
+  if (isType(action, removeToastr)) {
+    return {
+      ...state,
+      notifications: state.notifications.filter(x => x.id !== action.payload.id)
+    };
+  }
+  return state;
+}
