--- front/src/reducers/environmentConfig.ts
@@ -1,33 +1,0 @@
-/*
- * Copyright (c) 2018 Works Applications Co., Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-import { Action } from "redux";
-import { isType } from "typescript-fsa";
-import { getEnvironmentConfigsOperation } from "../actions";
-import { IEnvironmentConfigState } from "../store";
-
-const INIT: IEnvironmentConfigState = {
-  environmentConfigs: []
-};
-
-export function environmentConfig(
-  state: IEnvironmentConfigState = INIT,
-  action: Action
-): IEnvironmentConfigState {
-  if (isType(action, getEnvironmentConfigsOperation.async.done)) {
-    return { ...state, environmentConfigs: action.payload.result.configs };
-  }
-  return state;
-}
