--- front/src/reducers/index.ts
@@ -1,30 +1,0 @@
-/*
- * Copyright (c) 2018 Works Applications Co., Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-import { connectRouter } from "connected-react-router";
-import { combineReducers, Reducer } from "redux";
-import { IApplicationState } from "../store";
-import { analysisData } from "./analysisData";
-import { environmentConfig } from "./environmentConfig";
-import { toastNotification } from "./toastNotification";
-
-export default (history: any) =>
-  // tslint:disable-next-line:no-object-literal-type-assertion
-  combineReducers({
-    analysisData,
-    toastNotification,
-    environmentConfig,
-    router: connectRouter(history)
-  } as { [P in keyof IApplicationState]: Reducer<IApplicationState[P]> });
